backend/requirement.txt
paho-mqtt==2.1.0
requests==2.31.0
python-dotenv==1.0.1
Flask==3.0.2